% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aepet}
\alias{aepet}
\title{AEPET Data}
\format{A dataframe with 95 rows of 44 variables:}
\source{
Brain and Memory Lab
}
\usage{
aepet
}
\description{
This dataset contains various behavioral and personaltiy measures from the AEPET study.
}
\keyword{datasets}
