% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{corr_cv}
\alias{corr_cv}
\title{Cross-Validated Correlation}
\usage{
corr_cv(x, y)
}
\arguments{
\item{x}{A single vector.}

\item{y}{A single vector.}
}
\value{
An average correlation coefficient and the p-value for that coefficient.
}
\description{
This function allows you to run a correlation using leave-one-out cross-validation.
}
\examples{
x <- rnorm(n = 25, mean = 0, sd = 1)
y <- rnorm(n = 25, mean = 25, sd = 5)
corr_cv(x, y)
}
\keyword{correlation}
\keyword{cross-validation,}
