% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dicat}
\alias{dicat}
\title{DICAT Data}
\format{A dataframe with 38 rows of 31 variables:}
\source{
Brain and Memory Lab
}
\usage{
dicat
}
\description{
DICAT is a really boring category learning task. The behavioral results are not included in this dataset.
This dataset contains the correlations between the prototype / exemplar model fits and certain brain regions,
during final test runs 1 - 4 and the average correlations across the 4 test runs.
}
\keyword{datasets}
