% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{facat}
\alias{facat}
\title{FACAT Data}
\format{A dataframe with 29 rows of 73 variables:}
\source{
Brain and Memory Lab
}
\usage{
facat
}
\description{
This dataset contains various behavioral and fMRI measures from the FACAT study. 
FACAT is a category learning in which participants learned the underlying
category structure of blended faces. Some faces shared perceptual features - i.e. 50% blends - that
were relevant to the category structure, some faces shared perceptual features that were
irrelevant to the category structure, some faces shared no perceptual features. Included in the
dataset are the results from a representational similarity analysis looking at the item-level representations
and category-level representations.
}
\keyword{datasets}
