% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{groupsum}
\alias{groupsum}
\title{Group Summary Statistics}
\usage{
groupsum(data, y, g1, g2 = NULL, g3 = NULL, g4 = NULL)
}
\arguments{
\item{data}{A dataframe that contains the variables to be summarized.}

\item{y}{A column from the dataframe specified that will be summarized as an outcome measure.}

\item{g1}{Grouping variable 1}

\item{g2}{Grouping variable 2}
}
\value{
A dataframe with the mean and standard error by grouping factor(s).
}
\description{
This function summarizes your data by a given factor(s). It will return a dataframe that includes
the mean and standard error of the outcome measure of interest. The summary statistics will be 
formatted for plotting using ggplot2.
}
\examples{
dime <- dime 
groupsum(data = dime, y = acc, g1 = modul)
}
\keyword{error,}
\keyword{mean,}
\keyword{standard}
\keyword{summarize,}
\keyword{tidy}
